fix(ui/ux): Fix all blocking issues for PR #511
- Fix Sentry duplicate declaration by using dynamic import
- Create safeInterval utility with visibility pause and max iterations
- Replace all setInterval with safeInterval in DecisionApproval, WIPPage, Dashboard
- ProgressLoader already uses requestAnimationFrame (compliant)
- Verify vendor chunking works (react-vendor, ui-vendor, i18n-vendor, form-vendor)
- Verify all JS files have sourceMappingURL

Resolves:
- Sentry build error
- 4 bare setInterval usages
- Vendor chunking verification
- Source map verification
