+++ handoff/20250928/40_App/frontend-dashboard/src/components/DecisionApproval.jsx
@@ -17,6 +17,7 @@ import { Label } from '@/components/ui/label'
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog'
 import { Progress } from '@/components/ui/progress'
 import { useToast } from '@/hooks/use-toast'
+import { safeInterval } from '@/lib/safeInterval'
 
 const DecisionApproval = () => {
   const { toast } = useToast()
@@ -116,33 +117,17 @@ const DecisionApproval = () => {
   const [approvalComment, setApprovalComment] = useState('')
 
   useEffect(() => {
-    // 模擬自動倒計時 - 有限次數 + 可見性暫停
-    let n = 0, id
-    const step = () => {
+    // 模擬自動倒計時 - 使用 safeInterval
+    const cleanup = safeInterval(() => {
       setPendingDecisions(prev => 
         prev.map(decision => ({
           ...decision,
           auto_approve_in: Math.max(0, decision.auto_approve_in - 1)
         }))
       )
-      if (++n >= 120) {
-        clearInterval(id)
-      }
-    }
-    const vis = () => {
-      if (document.hidden) {
-        clearInterval(id)
-      } else {
-        id = setInterval(step, 1000)
-      }
-    }
-    document.addEventListener("visibilitychange", vis)
-    vis()
+    }, 1000, 120)
 
-    return () => {
-      clearInterval(id)
-      document.removeEventListener("visibilitychange", vis)
-    }
+    return cleanup
   }, [])
 
   const handleApprove = async (decisionId, comment = '') => {
